requirements.txt
pytest~=8.3.5
pytest-html==4.1.1
pytest-xdist==3.5.0
selenium==4.18.1
Appium-Python-Client~=5.0.0
requests==2.31.0

allure-python-commons~=2.13.5
PyYAML~=6.0.2